gym_agent/requirements.txt
langgraph>=0.0.30
openai>=1.12.0
qdrant-client>=1.7.0
meilisearch>=0.25.0
mcp>=1.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
tavily-python>=0.2.2
fastapi>=0.104.1
uvicorn>=0.24.0 